Finalize first test based on description - Context
We only have to write the code that decides which seats to suggest, so
we have all the elements ready and can start testing the simplest case:
suggest the only seat available.

For that, we need a seat to test the result, so we create one.

Of course, the system, cannot return it if it does not know it, so we
have to provide the seat to the system. It is not part of the requested
API (only the price range is requested), so we assume the seats to be
known from the start and ready to pick: we provide it through the
constructor rather than an additional method argument.

Motivating this design choice is a matter of conception (ontology of the
client we work for). Ontologically (as per the consensus established
with the client), we consider that the concept of the system makes no
sense without the seats on which it operates. The seats are thus a
mandatory attribute of the system, so the latter cannot be created
without the former.

We implement what we need to make the test pass.