Finalize first test based on description - Feature
The suggestion operates based on a price range, which we didn't used
yet. As per the description, the price range is a filter on the seat, so
for the test to make sense, the seat must comply to the price range.

We add the required methods and implementations to represent this
relation without breaking the test.