Finalize first test based on description - Explicitness
Reading the test, we can see that it remains mysterious: why would the
seat be returned based on a price range which does not show clearly that
the seat we provide fits?

The requirement is ill-defined by the test!

We want to ensure that the seat fits the provided price range. So we
must define a seat price and a price range which are compatible within
the test. The simplest case is the price range which exactly matches the
price.

The test already passes, so we only need to adapt the design of our
classes, without changing the implementation.