Test extension - Seat number
HAving one seat is nice, but makes the suggestion a bit too obvious. We
want to support 2 seats now. Unfortunately our design does not allow
that. So let's generalize our design to accept more seats.

Since we decide to change the type of the field, we need to adapt the
algorithm to compile... The test passes, so that will be totally fine.