Test generalization - Refactor seats
OK, we can see that coming: 3, 4, 5, 10, 100 seats, that will be boring
to read, annoying to write, and hard to maintain. We just change the
number of seats and expect them to be all in the suggestion anyway. So
let's aim for a parameterized test, so we can focus on the test cases
while reusing the test itself.

But the constructor is annoying with its varargs: we must provide an
array, and we don't want to have arrays everywhere. We are no C
programmer. So let's abstract that with a List, like the for the
assertion. This way we can just use the same for both.

We refactor both tests to ensure they remain equivalent in writing.