New test - Generalization
With just the new test, we already doubled the number of tests, but we
test only with the first seat taken as the target seat. Now we
generalize it to be able to use three variants for a better
representativeness: first, last, and middle seats.

We have now 20 tests in total. A good safety net before to start the
next requirement.