New test - not booked criteria
Let's ensure that we do not book twice a seat. However, as per the
description, we only have to implement the suggestion algorithm. So the
booking is managed by someone else. Thus, we only have to feed on some
external dependency to know whether a seat is free.

We abstract this dependency through a predicate, so any implementation
can be plugged to our system. It also makes it trivially testable.
Especially the existing cases, which just need a constant predicate to
work, which we name correctly to explicit its meaning.