New requirement - Adjacent seats
Let's go with the next requirement of having adjacent seats for the same
party.

From a lifecycle perspective, this information comes with the price
range: upon booking a seat, we select a price range and book the seats
of a party sequentially. It is not a global contextual data like the
available seats in the theater. Upon establishing the consensus with the
client, we confirm that the API has to be changed to take the party's
seats (already booked) as an additional parameter.

We adapt the existing tests correspondingly.