New test - Adjacent seat of party first 2
There is so much to complain about the previous test, but let's take it
one at a time.

First, the case chosen is just luckily compatible with the current code.
Let's add the same test but with reversed order of seats: the party
books the last seat, so we expect the 2nd then the 1st.

Of course it does not pass, the order being reversed. We need to check
the distance with the party's seat to properly order them.

The distance calculation is a constant: the distance between 2 seats
does not change in time. Thus it has the same lifetime as the set of all
seats. We establish this consensus with the client, as well as the fact
that we count the distance in number of seats (integer), and set it as a
constructor argument.

We set the simple distance computation for both tests and use a trivial
one for the other existing tests to not affect their order (all seats
are at the same distance).

The implementation is of course changed to make them pass.