New test - Adjacent seat of party first 3
Now we want a test for the middle case. For this one, we expect both
adjacent seats to be at the top. For 3 seats, the test might be a little
uninformative, because it returns all the non-booked seats in any order.
So let's try with 5 seats: we should have the 3rd seat booked for the
party, and the suggested seats starting with 2nd and 4th seats, in any
order.

Now we get a problem, because we need the suggested seats to be ordered
for retrieving the first ones. So the method should not return a
Collection but a List. We can keep the other tests with their Collection
result, though, since they don't need more.

Although we need the suggestion to be ordered, the comparison needs to
be unordered, because we don't care which of the adjacent seats come
first, as long as they all come together. Since the seats should not
come multiple times in a single list, we assume that using a set of them
allows to compare them in an unordered manner. We extract that in a
dedicated assert methods to hide this implementation detail.