Refactor - Topology-generic
One important thing of the current design is that it is totally
oblivious to the topology of the seats. They might be in row, column,
square, cube, sphere or whatever. We just get a list of seats and the
distance between them to know how to order them.

We can however take a Collection instead of a List to generalize more
and show that we don't care about the order of the seats.