New test - Adjacent seat of party first 4
The fact it is oblivious to the topology means that we don't need to
test for a 2D topology to ensure that the feature works in 2D. However,
focusing on 1D only might not bring us the relevant corner cases.

To cover a more generic setting and ensuring that the key criteria here
is the adjacency, we can go for an arbitrary collection of seats with an
arbitrary topology, where an arbitrary half of the seats are adjacent to
the party seat. This way, we can check that the order properly follow
the same arbitrariness of the context.

In general, this distance is not even Euclidean, so it makes our feature
rather flexible.