New test - Nearest to middle of row
Again, we want to prioritize based on a distance. Again it appears to be
counted in number of seats, with the middle seat of a row being at
distance 0.

However, we are not comparing two seats. We compare 1 seat to something
which can be derived from it: from the seat we get the row it is at, and
from the row we get the middle of it. For a row with an odd number of
seats, we would compare two seats, but with an even number of seats, the
middle would be either between two seats or the two seats themselves,
not a single seat. So we cannot reuse the same definition that we used
for the seat distance earlier (comparing two seats). Here we take a
single seat, while the second element has to be retrieved from the
first.

Yet, the remaining thinking is similar to the previous distance, so we
integrate it the same way to make the new test pass (distance in
constructor argument + comparator based on it).

We test it in the simple case of a single row of 3 seats, which should
returns the 2nd seat first.