New test - party > row middle
Now that we have tested separately the impact of the different distances
(party, row middle, and stage), we can ensure that the priority
described in the requirements is fulfilled.

We can test it in 2 ways.
- 1 complete priority test: all the distances are defined together on a
test case which contains conflicting distances (priorize seat A over
distance X, but seat B over distance Y) for each relevant distance
combination. The difficulty is to build a test case which produces all
the relevant conflicts, which can make it rather complex and will
require to change when the reauirements change as well.
- 1 priority test per pair: we test the priority between 2 distances at
a time, and repeat the operation for all the relevant combinations. It
requires more tests, but each test remain simple and changing the
requirements simply results in discarding existing tests or writing new
ones.

For maintainability purpose, we go through the second way (we want the
next people who will work on our code to not struggle with it).

Of course, we can also go the hybrid route by making a single test for
few compatible distances (row middle + stage complement each other
rather well to be tested together), but let's take the extreme route
here.

So we write a single test to priorize party distance over the middle of
the row.