New test - row middle > stage
Now we test the priority of the ,iddle of the row over the stage
distance.

We assume a transitivity of the priorities, so we will not test the
priority of the party over the stage and assume that it is already
covered by the previous tests. But it would be a trivial matter to do
so.