Final refactoring - Remove useless code
Based on test coverage, this code is not touched. The general process it
to clarify whether it should have its use or not. If yes, we should have
some tests covering it (might be by testing other classes). If not, we
should get rid of it (YAGNI). A deeper thinking about
defensive/offensive programming might bring more nuances to this
dichotomy, though.

Here we go the YAGNI route, because in this exercise we only care about
the suggestions. In a real case, it would surely go differently.

Anyway, now we are at 100% of branch coverage.